Only elect stateful process classes
